CartDiscountDraft to interface. #942
